==> 2_child_adult/analysis/6_bin_outliers.Rmd <==
---
title: "6_bin_outliers"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# load packages
```{r message=FALSE, warning=FALSE}
library(tidyverse)
library(here)
library(janitor)
```


# read in combined clean no BL outliers data 
```{r}
bins <- read_csv(here::here("data", "adult_child_combined", "zdiff_cheek_brow.csv"))
```


# make new screening column 

... which is TRUE if the zdiff value for a particular bin is greater than 2sd above the mean zdiff value for tha trial for that participant/emotion/muscle
```{r}
bins_out <- bins %>%
  group_by(pp_no, model, emotion, muscle, trial) %>%
  mutate(meanZdiff = mean(zdiff), sdZdiff = sd(zdiff)) %>%
  mutate(bin_outlier = if_else(zdiff > meanZdiff + 2*sdZdiff, "TRUE", "FALSE"))
```

Make sure the new screen column is logical TRUE FALSE
```{r}
bins_out$bin_outlier <- as.logical(bins_out$bin_outlier)
```

How many bins are bad?
```{r}
outlier_true <- bins_out %>%
  filter(bin_outlier == TRUE)
```

# test on 1 participant, 1 trial 

pick a pp that has an example of bad bins and work out how to make the bad bin value NA
```{r}
example <- bins_out %>%
  filter(pp_no == "pp406") %>%
  filter(emotion == "121") %>%
  filter(trial == "trial1")

glimpse(example)
```
# use mutate and replace to replace the Zdiff value with NA

...if bin_outlier -- TRUE

```{r}
test <- example %>%
  mutate(zdiff = replace(zdiff, bin_outlier == "TRUE", NA))
```

# apply bin screen to whole dataframe
```{r}
bins_screen <- bins_out %>%
  mutate(zdiff = replace(zdiff, bin_outlier == "TRUE", NA))
```

# drop extra columns, write csv
```{r}
bins_screen <- bins_screen %>%
  select(1:7) %>%
  write_csv(here::here("data", "adult_child_combined", "zdiff_binscreened.csv"))
```
